MAGE-986 Move new config options to discrete Autocomplete config helper
