+++ Block/Algolia.php
@@ -4,6 +4,7 @@ namespace Algolia\AlgoliaSearch\Block;
 
 use Algolia\AlgoliaSearch\Helper\AlgoliaHelper;
 use Algolia\AlgoliaSearch\Helper\ConfigHelper;
+use Algolia\AlgoliaSearch\Helper\Configuration\AutocompleteHelper;
 use Algolia\AlgoliaSearch\Helper\Configuration\PersonalizationHelper;
 use Algolia\AlgoliaSearch\Helper\Data as CoreHelper;
 use Algolia\AlgoliaSearch\Helper\Entity\CategoryHelper;
@@ -40,6 +41,8 @@ class Algolia extends Template implements CollectionDataSourceInterface
 
     public function __construct(
         protected ConfigHelper          $config,
+        protected AutocompleteHelper    $autocompleteConfig,
+        protected PersonalizationHelper $personalizationHelper,
         protected CatalogSearchHelper   $catalogSearchHelper,
         protected ProductHelper         $productHelper,
         protected Currency              $currency,
@@ -53,7 +56,6 @@ class Algolia extends Template implements CollectionDataSourceInterface
         protected CategoryHelper        $categoryHelper,
         protected SuggestionHelper      $suggestionHelper,
         protected LandingPageHelper     $landingPageHelper,
-        protected PersonalizationHelper $personalizationHelper,
         protected CheckoutSession       $checkoutSession,
         protected DateTime              $date,
         protected CurrentCategory       $currentCategory,
